tools: codestyle fixes, formatting updates to core/router/apps

+++ apps/addressbook/java/test/junit/net/i2p/router/naming/BlockfileNamingServiceTest.java
@@ -1,5 +1,10 @@
 package net.i2p.router.naming;
 
+import junit.framework.TestCase;
+
+import net.i2p.I2PAppContext;
+import net.i2p.data.DataHelper;
+
 import java.io.BufferedOutputStream;
 import java.io.File;
 import java.io.FileOutputStream;
@@ -10,9 +15,6 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Properties;
 import java.util.Set;
-import junit.framework.TestCase;
-import net.i2p.I2PAppContext;
-import net.i2p.data.DataHelper;
 
 @SuppressWarnings("PMD.CloseResource")
 public class BlockfileNamingServiceTest extends TestCase {
@@ -20,7 +22,7 @@ public class BlockfileNamingServiceTest extends TestCase {
     List<String> _names;
     File hostsTxt, routerDir;
 
-    @SuppressWarnings({ "unchecked", "rawtypes" })
+    @SuppressWarnings({"unchecked", "rawtypes"})
     public void setUp() throws Exception {
         I2PAppContext ctx = new I2PAppContext();
         routerDir = ctx.getRouterDir();
@@ -28,9 +30,9 @@ public class BlockfileNamingServiceTest extends TestCase {
         // first load the list of hosts that will be queried
         InputStream is = getClass().getResourceAsStream("/hosts.txt");
         Properties props = new Properties();
-        assertNotNull("test classpath not set correctly",is);
+        assertNotNull("test classpath not set correctly", is);
         DataHelper.loadProps(props, is, true);
-        _names = new ArrayList<String>((Set<String>) (Set) props.keySet());  // TODO-Java6: s/keySet()/stringPropertyNames()/
+        _names = new ArrayList<String>((Set<String>) (Set) props.keySet()); // TODO-Java6: s/keySet()/stringPropertyNames()/
         Collections.shuffle(_names);
         is.close();
 
@@ -38,22 +40,21 @@ public class BlockfileNamingServiceTest extends TestCase {
         hostsTxt = new File(routerDir, "hosts.txt");
         OutputStream os = new BufferedOutputStream(new FileOutputStream(hostsTxt));
         is = getClass().getResourceAsStream("/hosts.txt");
-        byte [] b = new byte[8196];
+        byte[] b = new byte[8196];
         int read = 0;
-        while ((read = is.read(b)) > 0)
-            os.write(b,0,read);
-        os.flush(); os.close();
+        while ((read = is.read(b)) > 0) os.write(b, 0, read);
+        os.flush();
+        os.close();
         _bns = new BlockfileNamingService(ctx);
     }
 
     public void tearDown() {
         _bns.shutdown();
         if (routerDir != null) {
-            File f = new File(routerDir,"hostsdb.blockfile");
+            File f = new File(routerDir, "hostsdb.blockfile");
             f.delete();
         }
-        if (hostsTxt != null)
-            hostsTxt.delete();
+        if (hostsTxt != null) hostsTxt.delete();
     }
 
     public void testRepeatedLookup() throws Exception {
